Refactor to inject Strategy
