+++ experiments/mock-experiment/experiment/__init__.py
@@ -2,6 +2,7 @@ from experiment.experiment import experiment
 from experiment.model import model
 from experiment.flower_client import FlowerClient
 from experiment.get_eval_fn import eval_fn
+from experiment.strategy import Strategy
 import experiment.settings as settings
 
-__all__ = ["experiment", "model", "FlowerClient", "eval_fn", "settings"]
+__all__ = ["experiment", "model", "FlowerClient", "eval_fn", "Strategy", "settings"]
