+++ experiments/mock-experiment/experiment/strategy.py
@@ -1,0 +1,4 @@
+import flwr
+
+class Strategy(flwr.server.strategy.Strategy):
+    pass
