+++ services/fl-server/requirements.txt
@@ -1,2 +1,1 @@
 flwr
-numpy
